requirements.txt
coverage>=4.5.2
matplotlib>=3.1.0
pandas>=0.24.2
Pillow>=6.0.0
pylint>=1.7.4
pytest>=3.2.3
torch>=1.1.0
torchvision>=0.3.0